apps/workers/pyproject.toml
[tool.poetry]
name = "workers"
version = "0.1.0"
description = "Worker services for AI Spreadsheet Assistant"
authors = ["Development Team"]

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pydantic = "^2.11.3"
supabase = "^2.4.0"
httpx = ">=0.24.0,<0.26.0"
redis = "^5.0.1"
celery = "^5.3.6"
sentry-sdk = "^1.44.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
